Tests/LibWeb/Text/expected/css/scrollable-overflow-recalc-animation.txt

initial scrollHeight: 100
scrollHeight at animation end: 550
scrollHeight at animation start: 100
relayouts performed: 0
